+++ R/utils.R
@@ -105,3 +105,7 @@ vec_is_coercible <- function(x, to, ..., x_arg = "x", to_arg = "to") {
 last <- function(x) {
   x[[length(x)]]
 }
+
+str_compact <- function(x) {
+  x[x != ""]
+}