+++ tests/testthat/test-inds-combine.R
@@ -51,8 +51,11 @@ test_that("if multiple names, last kept", {
   expect_identical(wrn$var, c(d = 1L, e = 1L))
 })
 
-test_that("if one name for multiple vars, use integer index", {
-  expect_equal(inds_combine(letters[1:3], list(x = 1:3)), c(x1 = 1, x2 = 2, x3 = 3))
+test_that("combine names if one name for multiple vars", {
+  expect_identical(
+    inds_combine(letters[1:3], list(x = 1:3)),
+    c(x1 = 1L, x2 = 2L, x3 = 3L)
+  )
 })
 
 test_that("select(0) corner case #82", {