+++ NEWS.md
@@ -43,6 +43,12 @@
   only the last one is taken into account (#52). The warning inherits
   from `tidyselect_warning_duplicate_renaming`.
 
+* `vars_select()` now ignores existing duplicate variables (#94). It
+  fails if user attempts to rename to an existing variable, or if
+  different variables are renamed to the same name. The errors inherit
+  from `tidyselect_error_rename_to_existing` or
+  `tidyselect_error_rename_to_same` respectively.
+
 * `one_of()` now always coerces its input to a character, or fails. Similary,
   `vars_select()` now supports unquoting S3 vectors.